Refactor: Decouple Agent from Proxy TLS dependency
