+++ pkg/core/certs/asset/setup_ca.sh
